requirements.txt
asgiref==3.4.1
Django==3.2.5
pytz==2021.1
sqlparse==0.4.1
requests
attrs
redis
channels
channels_redis
six
websocket-client
websockets